backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.8
Django>=5.1.1
djangorestframework>=3.12.0,<4.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=3.7.0
filelock==3.13.4
fonttools==4.53.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.1
psycopg2==2.9.9
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
resend==2.4.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.0
whitenoise>=5.3.0
psycopg2-binary>=2.9.1
python-decouple>=3.6
django-debug-toolbar>=3.2.1
django-crontab>=0.7.1
gunicorn>=20.1.0


